requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.23.0
lapx>=0.5.2
python-dotenv>=1.0.0
requests>=2.28.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
websockets>=11.0
python-multipart>=0.0.6
